requirements.txt
Flask==2.3.3
chromadb==0.5.0
python-docx==1.1.0
google-generativeai==0.7.0
gunicorn==22.0.0